+++ src/new/ctl/new/ctl/math/math.hpp
@@ -4,6 +4,5 @@
 #include "../cmath.hpp"
 #include "core.hpp"
 #include "p3.hpp"
-#include "random.hpp"
 
 #endif // CTL_MATH_H
