+++ src/new/ctl/new/ctl/random/random.hpp
@@ -1,0 +1,7 @@
+#ifndef CTL_RANDOM_H
+#define CTL_RANDOM_H
+
+#include "engine.hpp"
+#include "generator.hpp"
+
+#endif
